simplify and document minio tests
