+++ packages/@uppy/aws-s3/vitest.config.ts
@@ -3,21 +3,19 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     testTimeout: 120_000,
-    globalSetup: ['./tests/s3-client/setup.js'],
-    setupFiles: ['./test-setup.mjs'],
+    globalSetup: ['tests/s3-client/setup.ts'],
     projects: [
       {
         test: {
-          name: 'aws-s3-node',
-          include: ['tests/index.test.ts', 'tests/createSignedURL.test.ts'],
+          name: 's3-jsdom',
+          include: ['**/*.test.ts'],
           environment: 'jsdom',
         },
       },
       {
         test: {
-          name: 's3-client-browser',
-          include: ['tests/s3-client/*.test.{js,ts}'],
-          exclude: ['tests/s3-client/_shared.test.js'],
+          name: 's3-browser',
+          include: ['**/minio.test.ts'],
           browser: {
             enabled: true,
             provider: 'playwright',
